+++ go/2022/day05/instruction.go
@@ -1,5 +1,5 @@
 package main
 
 type Instruction struct {
-	count, from, to int64
+	count, from, to int
 }
